Big clean up
