+++ .vscode/settings.json
@@ -12,6 +12,7 @@
         ".\\Cargo.toml",
         ".\\Cargo.toml",
         ".\\Cargo.toml",
+        ".\\Cargo.toml",
         ".\\Cargo.toml"
     ],
     "rust-analyzer.showUnlinkedFileNotification": true,
