+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sr-rl"
-version = "0.19.0"
+version = "0.20.0"
 edition = "2021"
 rust-version = "1.76"
