+++ src/Engine/render/mod.rs
@@ -2,21 +2,18 @@
 
 use bevy::prelude::*;
 
-pub mod tilemap_render;
-pub mod pieces_render;
 pub mod components;
-pub mod cursor_render;
+mod tilemap_render;
+mod pieces_render;
+mod cursor_render;
 
+
+use crate::game::states::GameState;
 
 use self::{
     tilemap_render::spawn_map_render,
     pieces_render::{spawn_piece_renderer, spawn_exit_render}, //melee_animation
     cursor_render::spawn_game_cursor,
-};
-
-use crate::{
-    globals::STANDARD_TILE_SIZE, 
-    game::states::GameState, vectors::Vector2Int, 
 };
 
 pub struct GraphicsPlugin;
@@ -28,10 +25,6 @@ impl Plugin for GraphicsPlugin {
             .add_systems(OnEnter(GameState::Initialise), spawn_piece_renderer)
             .add_systems(OnEnter(GameState::Initialise), spawn_game_cursor)     
             .add_systems(OnEnter(GameState::Initialise), spawn_exit_render)    
-
-
-
-
             // La première camera.
             .add_systems(Startup, spawn_camera)    
             ;
@@ -61,18 +54,3 @@ fn spawn_camera(mut commands: Commands) {
 
 
 
-// TODO : Changer de place?
-
-pub fn get_world_position(
-    v: &Vector2Int
-) -> (f32, f32) {
-        // REMEMBER : Y in bevy2d = Negative when going down!
-        let x = v.x * STANDARD_TILE_SIZE;
-        let y = v.y  * STANDARD_TILE_SIZE;
-
-        //println!("GetWorldPosition : {:?} gives {:?}. World position get grid position : {:?}", (v.x, v.y), (iso_x, iso_y), get_grid_position(iso_x as f32, 0.0 - iso_y as f32));
-
-        (x as f32,
-        0.0 - y as f32)     // REMEMBER : Y in bevy2d = Negative when going down!
-
-}
