+++ src/game/combat/mod.rs
@@ -48,10 +48,11 @@ use bevy::prelude::*;
 
 pub mod components;
 pub mod events;
-pub mod event_systems;  //TODO deplacer les elements publiques?
-pub mod rules;
-pub mod ia;
 pub mod action_infos;
+pub mod rules;
+mod ia;
+mod event_systems;
+
 
 
 use crate::{engine::animations::events::GraphicsWaitEvent, game::{
@@ -84,7 +85,7 @@ impl Plugin for CombatPlugin {
             .add_event::<RefreshActionCostEvent>()              // Recalcule le cout d'une action / deplacement.
             .add_event::<TickEvent>()                           // De retour en 0.19j : Donne le rythme en recheckant où en sont les acteurs du combat.
 
-            .add_event::<EntityEndTurnEvent>()         // Envoyé par l'Entité qui mets volontairement fin à son tour.    //TODO : Meilleur nom: c'est une Action d'un NPC. 
+            .add_event::<EntityEndTurnEvent>()         // Envoyé par l'Entité qui mets volontairement fin à son tour.   
   
             .configure_sets(Update, CombatSet::Logic)      
             .configure_sets(Update, CombatSet::Tick.after(CombatSet::Logic))
@@ -124,7 +125,7 @@ impl Plugin for CombatPlugin {
 }
 
 
-pub fn combat_clean_death(
+fn combat_clean_death(
     //mut commands: Commands,
     player_q: Query<&Player>,
     mut ev_message: EventWriter<MessageEvent>,   //NEW MESSAGE EVENT SYSTEM v0.15.2
@@ -150,6 +151,7 @@ fn tick(
 }
 
 /// Donne AP aux participants, créé le CombatInfos ressource, passe en StartTurn.
+/// // Est utilisé par un Message du Manager.
 pub fn combat_start(    
     mut commands: Commands,
     mut ev_newturn: EventWriter<CombatTurnStartEvent>,
@@ -168,7 +170,7 @@ pub fn combat_start(
 
 
 /// Ajoute les Participants du Turn au Combat dans la queue CombatTurnQueue.
-pub fn combat_turn_start(
+fn combat_turn_start(
     mut action_query: Query<(Entity, &mut ActionPoints)>,
     npc_query: Query<Entity, (With<ActionPoints>, Without<Player>)>,
     player_query: Query<Entity, (With<ActionPoints>, With<Player>)>,
@@ -202,7 +204,7 @@ pub fn combat_turn_start(
 
 
 /// On récupère le prochain combattant, puisque le précédent a fini.
-pub fn combat_turn_next_entity(
+fn combat_turn_next_entity(
     mut commands: Commands,
     mut queue: ResMut<CombatTurnQueue>,    
     action_points_q: Query<&ActionPoints>,
@@ -232,7 +234,7 @@ pub fn combat_turn_next_entity(
     ev_refresh_ap.send(RefreshActionCostEvent);
 }
 
-pub fn combat_turn_end(    
+fn combat_turn_end(    
     mut ev_newturn: EventWriter<CombatTurnStartEvent>,
     mut queue: ResMut<CombatTurnQueue>,
 ){
@@ -245,7 +247,7 @@ pub fn combat_turn_end(
 /// 0.19j c'est cette fonction qui donne le rythme ! REMEMBER => Elle est très importante.
 /// Regarde si tous les PA ont été dépensé par le personnage dont c'est le tour.
 /// Si c'est le cas, passe au perso suivant.
-pub fn combat_turn_entity_check(
+fn combat_turn_entity_check(
     mut commands: Commands,
     current_combat: ResMut<CombatInfos>,
     query_action_points: Query<(&ActionPoints, Option<&Player>, Option<&Frozen>)>,  // Frozen => entité qu'on ne veut pas utiliser car non active.
@@ -278,6 +280,7 @@ pub fn combat_turn_entity_check(
 }
 
 /// Retire les ActionPoints, Remove CombatInfos, change State.
+/// Sera utilisable par le Manager.
 pub fn combat_end(
     mut commands: Commands,
     fighters: Query<(Entity, &ActionPoints)>,
