+++ src/map_builders/components.rs
@@ -1,10 +1,6 @@
-use bevy::prelude::*;
-
-use super::map::Map;
-
-
+/* 
 #[derive(Resource)]
-pub struct MapGenHistory{
-    pub history: Vec<Map>,
-    pub index: usize,
-}
+struct MapGenHistory{
+    history: Vec<Map>,
+    index: usize,
+}*/
